old_requirements.txt



click==7.1.2
setuptools==62.2.0
pandas==1.4.2
#python-dateutil==2.8.2
numpy==1.22.4
#geopy>=2.2.0
Flask>=2.1.2
natsort==8.2.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

pip==22.3.1
wheel==0.38.4
Jinja2==3.1.2
packaging==23.1

pytest==6.2.4
responses==0.23.3

